==> 00-Brief-Tabular-Time-Series-In-R.Rmd <==
---
layout: post
title: "Time Series 00: Intro to Time Series Data in R - Managing Date/Time 
Formats & Simple Plots using ggplot2"
date:   2015-10-25
authors: [Megan A. Jones, Marisa Guarinello, Courtney Soderberg, Leah A. Wasser]
contributors: []
dateCreated:  2015-10-22
lastModified: `r format(Sys.time(), "%Y-%m-%d")`
packagesLibraries: [ggplot2]
categories: [self-paced-tutorial]
mainTag: tabular-time-series
tags: [time-series, phenology, R]
tutorialSeries: [tabular-time-series]
description: "This tutorial will demonstrate how to import a time series data 
set stored in .csv format into R. It will explore data classes and will walk
through how to convert date data, stored as a character string, into a date 
class that R can recognize and plot efficiently."
code1: /R/dc-tabular-time-series/00-Brief-Tabular-Time-Series-In-R.R
image:
  feature: NEONCarpentryHeader_2.png
  credit: A collaboration between the National Ecological Observatory Network (NEON) and Data Carpentry
  creditlink:
permalink: /R/brief-tabular-time-series-qplot/
comments: true
---

{% include _toc.html %}

## About
This tutorial will demonstrate how to import a time series data set stored in `.csv`
format into `R`. It will explore data classes for columns in a `data.frame` and 
will walk through how to 
convert a date, stored as a character string, into a date class that `R` can
recognize and plot efficiently.

**R Skill Level:** Intermediate - you've got the basics of `R` down.

<div id="objectives" markdown="1">

# Goals / Objectives
After completing this tutorial, you will:

  * Be able to open a `.csv` file in `R` using `read.csv()`and understand why we
  are using that file type.
  * Understand how to work data stored in different columns within a `data.frame` in `R`.
  * Understand how to examine `R` object structures and data `classes`.
  * Be able to convert dates, stored as a character class, into an `R` date 
  class.
  * Know how to create a quick plot of a time-series data set using `qplot`. 
  
## Things You’ll Need To Complete This Tutorial
You will need the most current version of R and, preferably, RStudio loaded on your computer to complete this tutorial.

### Install R Packages

* **ggplot2:** `install.packages("ggplot2")`

 [More on Packages in R - Adapted from Software Carpentry.]({{site.baseurl}}/R/Packages-In-R/)

## Download Data
{% include/dataSubsets/_data_Met-Time-Series.html %}

****

{% include/_greyBox-wd-rscript.html %}

</div>

## Data Related to Phenology
In this tutorial, we will explore atmospheric data (including temperature,
precipitation and other metrics) collected by sensors mounted on a 
<a href="http://www.neonscience.org/science-design/collection-methods/flux-tower-measurements" target="_blank">
flux tower</a> 
at the NEON Harvard Forest field site. We are interested in exploring 
changes in temperature, precipitation, PAR and day 
length throughout the year -- metrics that impact changes in the timing of plant 
<a href=" http://budburst.org/phenology_defined" target="_blank">phenophases (phenology)</a>. 


## About .csv Format
The data that we will use is in `.csv` (comma-separated values) file format. The
`.csv` format is a plain text format, where each value in the dataset is 
separate by a comma and each "row" in the dataset is separated by a line break. 
Plain text formats are ideal for working both across platforms (Mac, PC, LINUX, 
etc) and also can be read by many different tools. The plain text 
format is also less likely to become obsolete over time.

<i class="fa fa-star"></i>**Data Tip:**
<a href="https://en.wikipedia.org/wiki/Comma-separated_values" target="_blank"> 
For more on .csv format see this Wikipedia article.</a>
{: .notice}

## Import the Data 

To begin, let's import the data into `R`. We can use base `R` functionality
to import a `.csv` file. We will use the `ggplot2` package to plot our data. 

```{r load-packages}
# Load packages required for entire script. 
# library(PackageName)  # purpose of package
library(ggplot2)   # efficient, pretty plotting - required for qplot function

# set working directory to ensure R can find the file we wish to import
# setwd("working-dir-path-here")

```

<i class="fa fa-star"></i> **Data Tip:** Good coding practice -- install and 
load all libraries at top of script.
If you decide you need another package later on in the script, return to this
area and add it. That way, with a glance, you can see all packages used in a
given script. 
{: .notice }

Once our working directory is set, we can import the file using `read.csv()`. 

```{r import-csv}

# Load csv file of daily meteorological data from Harvard Forest
harMet.daily <- read.csv(
      file="NEON-DS-Met-Time-Series/HARV/FisherTower-Met/hf001-06-daily-m.csv",
      stringsAsFactors = FALSE
      )
```

### stringsAsFactors=FALSE
When reading in files we most often use `stringsAsFactors = FALSE`. This
setting ensures that non-numeric data (strings) are not converted to
factors. 

#### What Is A Factor?

A factor is similar to a category. However factors can be numerically interpreted 
(they can have an order) and may have a level associated with them. 

Examples of factors:

* **Month Names (an ordinal variable):** Month names are non-numerical but we know 
that April (month 4) comes after March (month 3) and each could be represented
by a number (4 & 3).
* **1 and 2s to represent male and female sex (a nominal variable):** Numerical
interpretation of non-numerical data but no order to the levels.  

 
<i class="fa fa-star"></i> **Data Tip:** <a href="http://www.stat.berkeley.edu/~s133/factors.html" target="_blank">
Read more about factors here.</a>
{: .notice }

After loading the data it is easy to convert any field that should be a factor by
using `as.factor()`. Therefore it is often best to read in a file with
`stringsAsFactors = FALSE`.

## Data.Frames in R
The `read.csv()` imports our `.csv` into a `data.frame` object in `R`. `data.frame`s 
are ideal for working with tabular data - they are similar to a spreadsheet.

``` {r dataframe}
# what type of R object is our imported data?
class(harMet.daily)

```

## Data Structure
Once the data are imported, we can explore their structure. There are several
ways to examine the structure of a data frame: 

* `head()`: shows us the first 6 rows of the data (`tail()` shows the last 6 
            rows). 
* `str()` : displays the structure of the data as `R` interprets it.

Let's use both to explore our data. 

```{r view-data-structure}

# view first 6 rows of the dataframe 
head(harMet.daily)

# View the structure (str) of the data 
str(harMet.daily)
```


<i class="fa fa-star"></i> **Data Tip:** You can adjust the number of rows
returned when using the `head()` and `tail()` functions. For example you can use
`head(harMet.daily, 10)` to display the first 10 rows of your data rather than 6.
{: .notice }

### Classes in R

The structure results above let us know that the attributes in our `data.frame`
are stored as several different data types or `classes` as follows:

* **chr - Character:** It holds strings that are composed of letters and
words. Character class data can not be interpreted numerically - that is to say
we can not perform math on these values even if they contain only numbers. 
* **int - Integer:**  It holds numbers that are whole integers without decimals.
Mathematical operations can be performed on integers.
* **num - Numeric:**  It accepts data that are a wide variety of numeric formats 
including decimals (floating point values) and integers. Numeric also accept 
larger numbers than **int** will.

Storing variables using different `classes` is a strategic decision by `R` (and 
other programming languages) that optimizes processing and storage. It allows:

* data to be processed more quickly & efficiently.
* the program (`R`) to minimize the storage size.

### Differences Between Classes 

Certain functions can be performed on certain data classes and not on others. 

For example:

```{r data-class-examples }

a <- "mouse"
b <- "sparrow"
class(a)
class(b)

# subract a-b 
a-b

```

You can not subtract two **character** values given they are not numbers.

```{r data-class-num }

c <- 2
d <- 1
class(c)
class(d)
# subract a-b 
c-d

```

Additionally, performing summary statistics and other calculations of different 
types of classes can yield different results. 

```{r data-class-example-2 }

# create a new object
speciesObserved <- c("speciesb","speciesc","speciesa")
speciesObserved

# determine the class
class(speciesObserved)

# calculate the minimum
min(speciesObserved)

# create numeric object
prec <- c(1,2,5,3,6)
# view class
class(prec)
# calculate min value
min(prec)

```

We can calculate the minimum value for `SpeciesObserved`, a **character** 
data class, however it does not return a quantitative minimum. It simply
looks for the first element, using alphabetical (rather than numeric) order. 
Yet, we can calculate the quantitative minimum value for `prec` a **numeric**
data class.  

## Plot Data Using qplot()

Now that we've got classes down, let's plot one of the metrics in our data, 
air temperature -- `airt`. Given this is a time series dataset, we want to plot
air temperature as it changes over time. We have a date-time column, `date`, so 
let's use that as our x-axis variable and `airt` as our y-axis variable.

We will use the `qplot()` (for *quick plot*) function in the `ggplot2` package.
The syntax for `qplot()` requires the x- and y-axis variables and then the R
object that the variables are stored in. 

<i class="fa fa-star"></i>**Data Tip:** Add a title to the plot using 
`main="Title string"`.
{: .notice }

```{r plot-data }
# quickly plot air temperature
qplot(x=date, y=airt, 
      data=harMet.daily,
      main="Daily Air Temperature\nNEON Harvard Forest Field Site")

```

We have successfully plotted some data. However, what is happening on the 
x-axis?

`R` is trying to plot EVERY date value in our data, on the x-axis. This makes it
hard to read. Why? Let's have a look at the class of the x-axis variable - date.

```{r view-class }
# View data class for each column that we wish to plot
class(harMet.daily$date)

class(harMet.daily$airt)

```

In this case, the `date` column is stored in our `data.frame` as a character
class. Because it is a character, `R` does not know how to plot the dates as a
continuous variable. Instead it tries to plot every date value as a text string. 
The `airt` data is numeric so that metric plots just fine.

## Date as a Date-Time Class 

We need to convert our `date` column, which is currently stored as a character 
to a `date-time` class that can be displayed as a continuous variable. Lucky 
for us, `R` has a `date` class.  We can convert the `date` field to a `date class`
using `as.Date()`.

```{r convert-date-time }

# convert column to date class
harMet.daily$date <- as.Date(harMet.daily$date)

# view R class of data
class(harMet.daily$date)

# view results
head(harMet.daily$date)
```

Now that we have adjusted the date, let's plot again. Notice that it plots
much more quickly now that R recognizes `date` as a date class. `R` can 
aggregate ticks on the x-axis by year instead of trying to plot every day!

```{r qplot-data}
# quickly plot the data and include a title using main=""
# In title string we can use '\n' to force the string to break onto a new line
qplot(x=date,y=airt, 
      data=harMet.daily,
      main="Daily Air Temperature w/ Date Assigned\nNEON Harvard Forest Field Site")  

```


<div id="challenge" markdown="1">
## Challenge: Using ggplot2's qplot function 
1. Create a quick plot of the precipitation. Use the full time frame of data available
in the `harMet.daily` object. 
2. Do precipitation and air temperature have similar annual patterns? 
3. Create a quick plot examining the relationship between air temperature and precipitation. 

Hint: you can modify the X and Y axis labels using `xlab="label text"` and 
`ylab="label text"`.

</div>

```{r challenge-code-plotting, echo=FALSE}
# 1
qplot(x=date,y=prec, 
      data=harMet.daily,
      main="Daily Precipitation \nNEON Harvard Forest",
      xlab="Date", ylab="Daily Precipitation")  

# 2 Precipiation does not appear to have the same striking annual pattern as 
# air temperature does. 

# 3
qplot(x=airt,y=prec, 
      data=harMet.daily,
      main="Relationship between precipitation & air temperature \nNEON Harvard Forest",
      xlab="Air Temperature", ylab="Daily Precipitation")  

```
