==> 07-Culmination-Work-With-NDVI-and-Met-Data-In-R.Rmd <==
---
layout: post
title: "Time Series Culmination Activity: Plot using Facets & Plot NDVI with Time Series Data"
date:   2015-10-18
authors: [Megan A. Jones, Leah A. Wasser]
contributors: []
dateCreated: 2015-10-22
lastModified: `r format(Sys.time(), "%Y-%m-%d")`
packagesLibraries: [ggplot2, scales, gridExtra, grid, dplyr, reshape2]
categories: [self-paced-tutorial]
mainTag: tabular-time-series
tags: [time-series, phenology, R]
tutorialSeries: [tabular-time-series]
description: "This tutorial is a data integration wrap-up culmination activity 
for the spatio-temporal time series tutorials."
code1: /R/dc-tabular-time-series/07-Culmination-Work-With-NDVI-and-Met-Data-In-R.R
image:
  feature: NEONCarpentryHeader_2.png
  credit: A collaboration between the National Ecological Observatory Network (NEON) and Data Carpentry
  creditlink: http://www.neoninc.org
permalink: /R/Work-With-NDVI-and-Met-Data-In-R/
comments: true
---

{% include _toc.html %}

## About
This tutorial is a culmination activity for the series on 
[working with tabular time series data in R ]({{ site.baseurl }}/tutorial-series/tabular-time-series/) 
and as part of a larger spatio-temporal tutorial series and Data Carpentry
workshop on 
[using spatio-temporal phenology-related data in R ]({{ site.baseurl }}/tutorial-series/neon-dc-phenology-series/).
The data used in this culmination activity has been extracted or previously
worked with in the tutorial series
[working with tabular time series data in R ]({{ site.baseurl }}/tutorial-series/tabular-time-series/)
and 
[working with raster time-series data in R ]({{ site.baseurl }}/tutorial-series/raster-time-series/).

**R Skill Level:** Advanced/ Intermediate - you've got the basics of `R` down
 and are comfortable using `ggplot2` and `dplyr`. 

<div id="objectives" markdown="1">

# Goals / Objectives
After completing this activity, you will:

 * have applied `ggplot2` and `dplyr` skills to a new data set.
 * learn how to set min/max axis values in `ggplot()` to align data on multiple plots. 

## Things You’ll Need To Complete This Tutorial
You will need the most current version of R and, preferably, RStudio loaded on
your computer to complete this tutorial.

### Install R Packages

* **ggplot2:** `install.packages("ggplot2")`
* **scales:** `install.packages("scales")`
* **gridExtra:** `install.packages("gridExtra")`
* **grid:** `install.packages("grid")`
* **dplyr:** `install.packages("dplyr")`
* **reshape2:** `install.packages("reshape2")`

[More on Packages in R - Adapted from Software Carpentry.]({{ site.baseurl }}/R/Packages-In-R/)

### Download Data 
{% include/dataSubsets/_data_Met-Time-Series.html %}

****

{% include/_greyBox-wd-rscript.html %}

****

### Recommended Tutorials
This tutorial relies on participants' familiarity with both `dplyr` and 
`ggplot2`. Prior to working through this culmination activity, we recommend the
following tutorials, if you are new to either of the `R` packages.

* [Subset & Manipulate Time Series Data with dplyr tutorial ]({{ site.baseurl }}/R/time-series-subset-dplyr/ "Learn dplyr") 

* [Plotting Time Series with ggplot in R tutorial ]({{ site.baseurl }}/R/time-series-plot-ggplot/ "Learn ggplot")  

</div>

## Plot NDVI & PAR using Daily Data

### NDVI Data
Normalized Difference Vegetation Index (NDVI) is an indicator of how green
vegetation is.  

<iframe width="560" height="315" src="https://www.youtube.com/embed/rxOMhQwApMc" frameborder="0" allowfullscreen></iframe>

NDVI is derived from remote sensing data based on a ratio the
reluctance of visible red spectra and near-infrared spectra.  The NDVI values
vary from -1.0 to 1.0.

The imagery data used to create this NDVI data were collected over the National
Ecological Observatory Network's
<a href="http://www.neonscience.org/science-design/field-sites/harvard-forest" target="_blank" >Harvard Forest</a>
field site. 

The imagery was created by the U.S. Geological Survey (USGS) using a 
<a href="http://eros.usgs.gov/#/Find_Data/Products_and_Data_Available/MSS" target="_blank" >  multispectral scanner</a>
on a
<a href="http://landsat.usgs.gov" target="_blank" > Landsat Satellite </a>.
The data files are Geographic Tagged Image-File Format (GeoTIFF). 
The tutorial 
[Extract NDVI Summary Values from a Raster Time Series]({{ site.baseurl }}/R/Extract-NDVI-From-Rasters-In-R/), 
explains how to create this NDVI file from raster data. 

### Read In the Data
We need to read in two datasets: the 2009-2011 micrometeorological data and the
2011 NDVI data for the Harvard Forest. 

```{r read-in-files}
# Remember it is good coding technique to add additional libraries to the top of
# your script 

library(lubridate) # for working with dates
library(ggplot2)  # for creating graphs
library(scales)   # to access breaks/formatting functions
library(gridExtra) # for arranging plots
library(grid)   # for arrangeing plots
library(dplyr)  # for subsetting by season

# set working directory to ensure R can find the file we wish to import
# setwd("working-dir-path-here")

# read in the Harvard micro-meteorological data; if you don't already have it
harMetDaily.09.11 <- read.csv(
  file="NEON-DS-Met-Time-Series/HARV/FisherTower-Met/Met_HARV_Daily_2009_2011.csv",
  stringsAsFactors = FALSE
  )

#check out the data
str(harMetDaily.09.11)

# read in the NDVI CSV data; if you dont' already have it 
NDVI.2011 <- read.csv(
  file="NEON-DS-Met-Time-Series/HARV/NDVI/meanNDVI_HARV_2011.csv", 
  stringsAsFactors = FALSE
  )

# check out the data
str(NDVI.2011)
```

In the NDVI dataset, we have the following variables:

* 'X': an integer identifying each row
* meanNDVI: the daily total NDVI for that area. (It is a mean of all pixels in
the original raster).
* site: "HARV" means all NDVI values are from the Harvard Forest
* year: "2011" all values are from 2011
* julianDay: the numeric day of the year
* Date: a date in format "YYYY-MM-DD"; currently in **chr** class

<div id="challenge" markdown="1">
## Challenge: Class Conversion & Subset by Time
The goal of this challenge is to get our data sets ready so that we can work 
with data from each, within the same plots or analyses.  

1. Ensure that date fields within both data sets are in the Date class. If not,
convert the data to the Date class. 

2. The NDVI data is limited to 2011, however, the meteorological data is from
2009-2011. Subset and retain only the 2011 meteorological data. Name it
`harMet.daily2011`.

HINT: If you are having trouble subsetting the data, refer back to
[Subset & Manipulate Time Series Data with dplyr tutorial]({{ site.baseurl }}/R/time-series-subset-dplyr/ "Learn dplyr")
</div>

``` {r challenge-code-convert-date, include=TRUE, results="hide", echo=FALSE}

# check to see class of date field
class(NDVI.2011$Date)
class(harMetDaily.09.11$date)

# convert chr class Date to date class Date
NDVI.2011$Date<- as.Date(NDVI.2011$Date)
harMetDaily.09.11$date<- as.Date(harMetDaily.09.11$date)

# confirm conversion
class(NDVI.2011$Date)
class(harMetDaily.09.11$date)

# 2
# Use dplyr to subset only 2011 data
harMet.daily2011 <- harMetDaily.09.11 %>% 
  mutate(year = year(date)) %>%   #need to create a year only column first
  filter(year == "2011")

# convert data from POSIX class to Date class; both "date" vars. now Date class
harMet.daily2011$date<-as.Date(harMet.daily2011$date)
```

Now that we have our data sets with Date class dates and limited to 2011, we can
begin working with both. 

## Plot NDVI Data from a .csv
These NDVI data were derived from a raster and are now integers in a
`data.frame`, therefore we can plot it like any of our other values using
`ggplot()`. Here we plot `meanNDVI` by `Date`.

``` {r plot-NDVI}
# plot NDVI by date
ggplot(NDVI.2011, aes(Date, meanNDVI))+
  geom_point(colour = "forestgreen", size = 4) +
  ggtitle("Daily NDVI at Harvard Forest, 2011")+
  theme(legend.position = "none",
        plot.title = element_text(lineheight=.8, face="bold",size = 20),
        text = element_text(size=20))

```

## Two y-axes or Side-by-Side Plots?
When we have different types of data like NDVI (scale: 0-1 index units),
Photosynthetically Active Radiation (PAR, scale: 0-65.8 mole per meter squared),
or temperature (scale: -20 to 30 C) that we want to plot over time, we cannot
simply plot them on the same plot as they have different y-axes.

One option, would be to plot both data types in the same plot space but each
having it's own axis (one on left of plot and one on right of plot).  However, 
there is a line of graphical representation thought that this is not a good
practice.  The creator of `ggplot2` ascribes to this dislike of different y-axes
and so neither `qplot` nor `ggplot` have this functionality. 

Instead, plots of different types of data can be plotted next to each other to 
allow for comparison.  Depending on how the plots are being viewed, they can
have a vertical or horizontal arrangement. 

<div id="challenge" markdown="1">
## Challenge: Plot Air Temperature and NDVI

Plot the NDVI vs Date (previous plot) and PAR vs Date (create a new plot) in the
same viewer so we can more easily compare them. 

Hint: If you are having a hard time arranging the plots in a single grid, refer
back to 
[Plotting Time Series with ggplot in R tutorial]({{ site.baseurl }}/R/time-series-plot-ggplot/ "Learn ggplot")
</div>

``` {r plot-PAR-NDVI, echo=FALSE}

# plot NDVI again
plot.NDVI.2011 <- ggplot(NDVI.2011, aes(Date, meanNDVI))+
  geom_point(colour = "forestgreen", size = 4) +
  ggtitle("Daily NDVI at Harvard Forest, 2011")+
  theme(legend.position = "none",
        plot.title = element_text(lineheight=.8, face="bold",size = 20),
        text = element_text(size=20))

# create plot of julian day vs. PAR
plot.par.2011 <- ggplot(harMet.daily2011, aes(date, part))+
  geom_point(na.rm=TRUE) +
  ggtitle("Daily PAR at Harvard Forest, 2011")+
  theme(legend.position = "none",
        plot.title = element_text(lineheight=.8, face="bold",size = 20),
        text = element_text(size=20))

# display the plots together
grid.arrange(plot.par.2011, plot.NDVI.2011) 
```

The figures from this Challenge are nice but a bit confusing as the dates on the
x-axis don't exactly line up. To fix this we can **assign the same min and max 
to both x-axes** so that they align. The syntax for this is: 

`limits=c(min=VALUE,max=VALUE)`. 

In our case we want the min and max values to 
be based on the min and max of the `NDVI.2011$Date` so we'll use a function 
specifying this instead of a single value.

We can also assign the date format for the x-axis and clearly label both axes. 

``` {r plot-same-xaxis}
# plot PAR
plot2.par.2011 <- plot.par.2011 +
               scale_x_date(labels = date_format("%b %d"),
               date_breaks = "3 months",
               date_minor_breaks= "1 week",
               limits=c(min=min(NDVI.2011$Date),max=max(NDVI.2011$Date))) +
               ylab("Total PAR") + xlab ("")

# plot NDVI
plot2.NDVI.2011 <- plot.NDVI.2011 +
               scale_x_date(labels = date_format("%b %d"),
               date_breaks = "3 months", 
               date_minor_breaks= "1 week",
               limits=c(min=min(NDVI.2011$Date),max=max(NDVI.2011$Date)))+
               ylab("Total NDVI") + xlab ("Date")

# Output with both plots
grid.arrange(plot2.par.2011, plot2.NDVI.2011) 

```

<div id="challenge" markdown="1">
## Challenge: Plot Air Temperature and NDVI
Create a plot, complementary to those above, showing air temperature (`airt`)
throughout 2011. Choose colors and symbols that show the data well. 

Second, plot PAR, air temperature and NDVI in a single pane for ease of
comparison.  
</div>

``` {r challengeplot-same-xaxis, echo=FALSE}
# 1
# plot air temp
plot.airt.2011 <- ggplot(harMet.daily2011, aes(date, airt))+
  geom_point(colour="darkblue", na.rm=TRUE) +
  ggtitle("Average Air Temperature\n Harvard Forest 2011")+
  scale_x_date(labels = date_format("%b %d"),
               date_breaks = "3 months", date_minor_breaks= "1 week",
               limits=c(min(NDVI.2011$Date),max=max(NDVI.2011$Date)))+
  ylab("Celcius") + xlab ("")+
  theme(legend.position = "none",
        plot.title = element_text(lineheight=.8, face="bold",size = 20),
        text = element_text(size=20))

plot.airt.2011

# 2 output with all 3 together
grid.arrange(plot2.par.2011, plot.airt.2011, plot2.NDVI.2011) 
```
